+++ plugin-go/goImpl/doc.go
@@ -1,2 +1,2 @@
-// Package goImpl provides the implementaiton of the zedpm-plugin-go.
+// Package goImpl provides the implementation of the zedpm-plugin-go.
 package goImpl