+++ plugin-golangci/golangciImpl/doc.go
@@ -1,0 +1,2 @@
+// Package golangciImpl provides the implementation of the zedpm-plugin-golangci.
+package golangciImpl