+++ plugin-golangci/zedpm-plugin-golangci.go
@@ -1,0 +1,10 @@
+package main
+
+import (
+	"github.com/zostay/zedpm/plugin-golangci/golangciImpl"
+	"github.com/zostay/zedpm/plugin/metal"
+)
+
+func main() {
+	metal.RunPlugin(&golangciImpl.Plugin{})
+}